pyproject.toml
[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[project]
name = "investments-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "jinja2>=3.1.6",
    "llama-cloud>=1.0.0b2",
    "llama-index-workflows>=2.12.0",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-multipart>=0.0.21",
    "tabulate>=0.9.0",
]

[tool.uv.build-backend]
module-name = "investments_review"

[project.scripts]
sheets-wf = "investments_review.sheets.main:main"
pres-wf = "investments_review.presentations.main:main"
serve = "investments_review.api:main"
